Hand-ported and extended the changes from the unfinished wbc-decruft-msg0 branch that got orphaned 3 months ago or so (but in wbcnet instead of wbcrun, because the latter will be phased out anyway).
wbcnet::msg::Service provides all the functionality we need for
actually implementing the user commands with the request/reply idea
that uses a code vector. Introduced the notion of "service domain"
which can be servo, behavior, task, or "other" (custom extensions) in
order to subdivide the commands into types based on what part of the
controller they interact with.

Also added the srv_command_t enum, which provides the second layer of
the command decoding protocol. So, actually, the code vector of a user
command request always has at least two elements: first the domain,
then the command.

Similarly to the extensions introduced for srv_result_to_string(), the
same happens for srv_domain_to_string() and srv_command_to_string().