Hooked some behavior commands back into the user process:  * "float" tells the behavior to go into float mode  * "activate" tells it to start doing "it's thing"  * "setgoal" sends a workspace goal for the end effector, directly    specified as 7 numbers (interpreted as x y z axis angle in Ellen's    WAM behaviors)  * "setgains" sends 4 numbers. Ellen's code interprets these as a type    code (1:x 2:y 3:yaw) followed by the values for kp, kd, and ki. Note that the setgoal and setgains commands are really quite specific to the WAM code.
Also re-patriated some of the commands from the SRV_SERVO_DOMAIN into
the SRV_BEHAVIOR_DOMAIN, because that's where they originally came
from. All legacy commands should probably belong to the behavior
domain, because historically that;s how they've been developed and if
we do not assign them to that domain, the command dispatcher will not
end up forwarding them to the currently running behavior.