Stuffed most things into a new attic:  * saimatrix because we really don't want to use it anymore  * wbc because we have to (i) redesign its foundation classes and (ii)    make it use jspace and Eigen2 instead of saimatrix  * wbc_plugins because it only makes sense within the old wbc    structure, where we needed to assure that closed-source binaries    can be loaded at runtime  * wbcnet because --- in spite of some useful and fairly generic    components --- it is used only within the narrow context of the    user/model/servo process split within the old wbc    applications. I'll copy-out generically useful things depending on    need, e.g. strutil (which comes from libsunflower.sf.net anyway).
The CMake setup now treats the remainder (tao, wbc_tinyxml, and
jspace) as non-standalone subprojects. The code of course still treats
each of them individually, and maybe one day we'll go back to
supporting individual build+install of each of them, but the simple
fact of the matter is that they only make sense together anyway.